fix error handling tests
