+++ app/controllers/admin_controller.rb
@@ -1,5 +1,6 @@
 class AdminController < ApplicationController
   include LoadTournament
+  include AdminErrorHandling
 
   abstract!
 
@@ -7,7 +8,6 @@ class AdminController < ApplicationController
 
   helper UiHelper
 
-  responders :flash
   respond_to :html
 
   before_action :load_tournament
@@ -16,31 +16,6 @@ class AdminController < ApplicationController
   before_action :set_tournament_cookie
 
   protected
-
-  unless Rails.application.config.consider_all_requests_local
-    rescue_from ActiveRecord::RecordNotFound do
-      render_admin_404
-    end
-
-    rescue_from Exception do |e|
-      Rollbar.error(e)
-      render_admin_500
-    end
-  end
-
-  def render_admin_404
-    respond_to do |format|
-      format.html { render 'admin/404', status: :not_found }
-      format.any  { head :not_found }
-    end
-  end
-
-  def render_admin_500
-    respond_to do |format|
-      format.html { render 'admin/500', status: 500 }
-      format.any  { head 500 }
-    end
-  end
 
   def authenticate_tournament_user!
     unless current_user.is_tournament_user?(@tournament.id) || current_user.staff?
